extract base RDFa parser class
added implementations with code specific to AngleSharp and HtmlAgilityPack
